+++ blocklets/ai-studio/src/locales/zh.ts
@@ -771,4 +771,5 @@ export default flatten({
   arrayItem: '数组项',
   integer: '整数',
   openEmbed: 'OpenEmbed',
+  selectAll: '全选',
 });